refactor: extracted methods to ExchangeRateUtils to reduce code duplicate
